test(logging): move sanitizer tests into mapped core-utils file
Codecov reported 75% patch coverage because the tests lived in
tests/logging_callback_tests/ (the Integrations shard, which does not
upload coverage for this source file). Move the 9 tests into
tests/test_litellm/litellm_core_utils/test_litellm_logging.py — the
mapped file for litellm_core_utils/litellm_logging.py, run by the
coverage-uploading core-utils shard — so patch coverage reflects them.
